% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_count_of.R
\name{add_count_of}
\alias{add_count_of}
\title{Find number of occurrences of a string in surveys}
\usage{
add_count_of(data, string)
}
\arguments{
\item{data}{dataframe containing survey data.}

\item{string}{character string that will be counted.}
}
\value{
A character vector having the count of occurrence of the string.
}
\description{
This function returns a vector containing the number of occurrences of a user-defined string (e.g. "Prefer not to answer") per survey throughout the dataframe.
The function takes as arguments a dataframe, the string looked for.

Note: the string needs to be harmonized across the dataframe. E.g. "don´t want to asnwer" should always be stored on the dataframe as "don´t want to answer".
}
\examples{
# Get the number of occurrences of "don´t want to answer" (stored as "dwta" on the dataframe).
\code{n} data <- data \%>\% mutate(check_nb_dwta = add_count_of(., "dwta"))

# Get the names of occurrences of "dont_know".
\code{n} data <- data \%>\% mutate(check_nb_dk = add_count_of(., "dont_know"))
}
