% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ratio_of.R
\name{add_ratio_of}
\alias{add_ratio_of}
\title{Find count ratio of an answer option throughout surveys}
\usage{
add_ratio_of(data, string, survey_df, choices_df)
}
\arguments{
\item{data}{dataframe containing survey data.}

\item{string}{character string that will be counted. Regex allowed.}

\item{survey_df}{dataframe containing ´survey´ sheet of the XLS form.}

\item{choices_df}{dataframe containing ´choices´ sheet of the XLS form.}
}
\value{
A numeric vector having the ratio(\%) of occurrence of the string.
}
\description{
This function returns a vector containing the count and ratio of a user-defined string (e.g. "other") per survey.
The function takes as arguments a dataframe, the string looked for, survey and choices sheets of the XLS forms.

Note: the string needs to be harmonized across the dataframe. E.g. "other" should always be stored on the dataframe as "other". Hoyever, you can use regex to find "other" and its alternatives e.g. "other_specify".
}
\examples{
# Get the ratio of "other" (stored as "other" on the dataframe).
\code{n} data <- data \%>\% mutate(check_others_ratio = add_ratio_of(., "dwta", kobo_survey, kobo_choices))

# Get the ratio of "other" and its alternatives.
\code{n} data <- data \%>\% mutate(check_others_etc_ratio = add_ratio_of(., "dont_know", kobo_survey, kobo_choices))
}
