% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_functions.R
\name{append_2change_logs}
\alias{append_2change_logs}
\title{Appends a change log to another}
\usage{
append_2change_logs(old, new, identifier = "identifier", q_colname = NULL)
}
\arguments{
\item{old}{dataframe, the main change/cleaning log.}

\item{new}{dataframe, the new log file to be appended to the main log.}

\item{identifier}{character, column name of the survey ID. Default: ´identifier´.}
}
\value{
A combined long table with distinct entries.
}
\description{
This function appends a change log to another while keeping distinct records.
}
\examples{
# Compiles two change/cleaning logs into one.
\code compiled_log<- append_2change_logs(old = cl_all, new = cl_new)
}
