% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_skip_logic_table.R
\name{apply_skip_logic}
\alias{apply_skip_logic}
\title{Applies skip logic to dataframe (Work in progress)}
\usage{
apply_skip_logic(data, skip_logic_table, formula_col)
}
\arguments{
\item{survey_df}{dataframe containing ´survey´ sheet of the XLS form.}
}
\value{
A list of two objects: 1. dataframe with skip logic applied. 2. change log
}
\description{
This function returns the new dataset with the skip logic applied to the columns.
NOTE: Ignores group skip logic. Only questions with question-level skip logic is considered.
}
\examples{
# Empties replies in certain questions based on skip logic.
\code data_cleaned <- apply_skip_logic(data, skip_logic_t, formula_col = "relevant_R")
}
