% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_security_checks.R
\name{check_FCS_all_same}
\alias{check_FCS_all_same}
\title{Checks if all values are the same for FCS}
\usage{
check_FCS_all_same(
  data,
  fcs_col_start = NULL,
  fcs_columns = NULL,
  ignore_condiment = TRUE
)
}
\arguments{
\item{data}{dataframe containing data.}

\item{fcs_col_start}{pattern to use for finding the FCS questions in the dataset.}

\item{fcs_columns}{vector containing column names of FCS questions.}

\item{ignore_condiment}{If FALSE, the condiment column will be ignored in this check. Defaults to TRUE.}
}
\value{
A numeric binary vector. 1 means the values are the same. 0 otherwise.
}
\description{
This function checks, per survey, if the values are the same across FCS questions.
}
\examples{
# Create a new column for the check as part of the dataset.
\code{n} data <- data \%>\% mutate(CHECK__FCS_same = check_FCS_all_same(data, fcs_col_start = "fs_fcs_", ignore_condiment = T))
}
