% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_security_checks.R
\name{check_FCS_alternating_values}
\alias{check_FCS_alternating_values}
\title{Checks by row if all values are the same in provided columns.}
\usage{
check_FCS_alternating_values(
  data,
  fcs_columns = NULL,
  fcs_col_pattern = NULL,
  n = 2,
  ignore_condiment = TRUE
)
}
\arguments{
\item{data}{dataframe. The dataset.}

\item{fcs_columns}{vector. The column names of FCS questions.}

\item{fcs_col_pattern}{character. The pattern to use for finding the FCS questions in the dataset.}

\item{ignore_condiment}{logical. If FALSE, the condiment column will be ignored in this check. Defaults to TRUE.}

\item{n_distinct}{numeric. the number of distinct digits to check for.}
}
\value{
A numeric binary vector. 1 means the values are the same. 0 otherwise.
}
\description{
This function checks, per row/survey, if the values are the same across the provided columns in the dataset.
Applicable to Food Consumption questions.
}
\examples{
# Create a new column for the check as part of the dataset.
\code data <- data \%>\% mutate(CHECK__FCS_alter = check_FCS_alternating_values(, fcs_col_pattern = "fs_fcs_", n = 2, ignore_condiment = T))
}
