% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_security_checks.R
\name{check_value_to_threshold}
\alias{check_value_to_threshold}
\title{Checks if a condition is respected in a column by comparing column values to a threshold.}
\usage{
check_value_to_threshold(
  data,
  column_to_check = NULL,
  operator = "<=",
  threshold = NULL
)
}
\arguments{
\item{data}{dataframe. The dataset.}

\item{column_to_check}{character. The column name of the column to check its values e.g. "fs_fcs_cerealgrainroottuber".}

\item{operator}{character. The mathematical operator to use. default: "<=".}

\item{threshold}{numeric. The threshold.}
}
\value{
A numeric binary vector. 1 means the specified condition is not met. 0 otherwise.
}
\description{
This function performs value checks. Right Hand Side (RHS) is the column, Left Hand side is for the threshold. If the condition is not met, it would be flagged as 1 in the result.
}
\examples{
# Create a new column for the check as part of the dataset.
\code data <- data \%>\% mutate(CHECK__FCS_cereals = check_value_to_threshold(data, column_to_check = "fs_fcs_cerealgrainroottuber", operator = "<=", threshold = 2))

\code data <- data \%>\% mutate(CHECK__FCS_cereals = check_value_to_threshold(., "fs_fcs_cerealgrainroottuber", "<=", 2))
}
