% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_skip_logic_table.R
\name{create_skip_logic_table}
\alias{create_skip_logic_table}
\title{Creates a skip logic dataframe from the XLS form}
\usage{
create_skip_logic_table(survey_df)
}
\arguments{
\item{survey_df}{dataframe containing ´survey´ sheet of the XLS form.}
}
\value{
A dataframe containing the names of the survey questions, their types, and their skip logic.
}
\description{
This function returns dataframe containing the question names, question type, and cleaned skip logic based on the the survey sheet of the XLS form.
NOTE: Only questions with skip logic are returned.
}
\examples{
# Creates input for cleaning up the questions based on skip logic.
\code columns_other <- find_columns_other(data, kobo_survey)
}
