% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_col_difference.R
\name{find_col_difference}
\alias{find_col_difference}
\title{Find difference between two submitted surveys}
\usage{
find_col_difference(data, identifier = "identifier", value, value2 = NULL)
}
\arguments{
\item{data}{dataframe containing survey data and unique identifier column for the surveys.}

\item{identifier}{column name of the survey ID. Default: ´identifier´.}

\item{value}{the survey ID in question. Used as a key for filtering out surveys for comparison.}

\item{value2}{optional parameter defining a different survey ID than the first. Useful for comparing surveys with different identifiers.}
}
\value{
A character string containing the names of the columns that are different. column names are separated with a comma ´,´ within the string.
}
\description{
This function returns a string containing all column names where values are different between surveys.
The function takes as arguments a dataframe, name of the unique identifier column, and 2 survey ID (1 mandatory, 1 optional).
}
\details{
Note: if you are comparing two surveys with the same identifier, only one identifier is needed.
}
\examples{
# Get the names of questions that have different responses on surveys. Fetching by 1 identifier i.e. surveys with duplicate identifier.
\code{n} names <- find_different_columns(data, "identifier", "3681f44c-8f1f-49b0-1dzc-6f741d087b24")

# Get the names of questions that have different responses on surveys. Fetching by 2 different identifiers
\code{n} names <- find_different_columns(data, "identifier", "3681f44c-8f1f-49b0-1dzc-6f741d087b24", "a211f44a-9d1f-94b0-10dc-6f741d811b29")
}
