% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_columns_other.R
\name{find_columns_other}
\alias{find_columns_other}
\title{Find column names of columns "other"}
\usage{
find_columns_other(data, survey_df)
}
\arguments{
\item{data}{dataframe containing data.}

\item{survey_df}{dataframe containing ´survey´ sheet of the XLS form.}
}
\value{
A character vector containing the names of the "other" variables.
}
\description{
This function returns a vector containing the names of columns that end with "other" and are text questions on the XLS form.
}
\examples{
# Get the column names of text "other" .
\code columns_other <- find_columns_other(data, kobo_survey)
}
