% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData function}
\usage{
formatData(
  dat,
  rescalev = TRUE,
  removepercentNAshigher = 1,
  cleanfromtxtconstants = TRUE,
  keepunscaled = NULL
)
}
\arguments{
\item{dat}{dataframe.}

\item{rescalev}{scales the columns of a numeric matrix (TRUE, by default)}

\item{removepercentNAshigher}{remove columns with percent of NAs exceeding given proportion (1, by default).}

\item{cleanfromtxtconstants}{remove columns with text/ constant columns percent of NAs exceeding given proportion (TRUE, by default).}

\item{keepunscaled}{vector of variables to keep  unscaled (NULL, by default).}
}
\value{
Returns the formatted data.
}
\description{
This function preformats the data.
}
\examples{
library(SurveyQuality)
library(dplyr)
dat <- read.csv(system.file("elites_survey.csv", package="SurveyQuality"))

dat[dat==98|dat==99|dat==95|dat==998|dat==999]<-NA
pmatch<-formatData(dat, rescalev=FALSE, removepercentNAshigher=0.90,
                   cleanfromtxtconstants=TRUE)\%>\%percentMatch()
}
