% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_functions.R
\name{generate_change_log}
\alias{generate_change_log}
\title{Generate change log with empty cleaning comments}
\usage{
generate_change_log(
  data1,
  data2,
  identifier = "identifier",
  data1_priority = FALSE
)
}
\arguments{
\item{data1}{dataframe. The first dataset, inclusive of unique identifier column.}

\item{data2}{dataframe. The second dataset, inclusive of unique identifier column.}

\item{identifier}{column name of the survey ID. Default: ´identifier´.}

\item{data1_priority}{Logical. only columns existent on data1 would be considered for both datatsets.}
}
\value{
A long table outlining each cell change (identified by the identifier) between the two datasets.
}
\description{
This function creates a change log by comparing two datasets.
}
\examples{
# Generate the cleaning log.
\code change_log <- generate_change_log(raw_data, clean_data, "uuid", data1_priority = T)
}
