% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interviewerEffects.R
\name{interviewerEffects}
\alias{interviewerEffects}
\title{interviewerEffects function}
\usage{
interviewerEffects(dat, f, separateffects = FALSE)
}
\arguments{
\item{dat}{dataframe}

\item{f}{lme formula}

\item{separateffects}{separate effects by group or not (FALSE, by default)}
}
\value{
Returns the list of results: the matrix of ICCs (intraclass correlation coefficients), the matrix of coefficients.
}
\description{
This function calculates the interviewer effects.
}
\examples{
library(SurveyQuality)
library(dplyr)
dat <- read.csv(system.file("elites_survey.csv", package="SurveyQuality"))

dat[dat==98|dat==99|dat==95|dat==998|dat==999]<-NA
dat$A6<-as.factor(dat$A6)
intEffects<-formatData(dat, rescalev=TRUE,
            removepercentNAshigher=0.80, cleanfromtxtconstants=TRUE)\%>\%
            interviewerEffects(f="y ~ 1 + A6 + (1|A5)")
intEffects$ICC
}
