% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessAnalysis.R
\name{missingnessAnalysis}
\alias{missingnessAnalysis}
\title{missingnessAnalysis function}
\usage{
missingnessAnalysis(dat)
}
\arguments{
\item{dat}{dataframe}
}
\value{
Returns the vector of item nonresponse rates for each interviewer.
}
\description{
This function computes an item nonresponse rate per interviewer.
}
\examples{
library(SurveyQuality)
library(dplyr)
dat <- read.csv(system.file("elites_survey.csv", package="SurveyQuality"))
dat[dat==98|dat==99|dat==95|dat==998|dat==999]<-NA
mistest<-missingnessAnalysis(dat)
mistest
}
