% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathR.R
\name{pathR}
\alias{pathR}
\title{Replaces backslash with slash in the provided file path}
\usage{
pathR(path)
}
\arguments{
\item{path}{Character, containing ´windows path´.}
}
\value{
A character string containing fixed file path.
}
\description{
\code{6} pathR function returns character string containing the file path fixed for use in R. It replaces backslash with slash.
NOTE: No checks are perfomed on the file path
}
\examples{
\code df <- pathR(path) \%>\% read_xlsx(., sheet = 1)
}
