% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentMatch.R
\name{percentMatch}
\alias{percentMatch}
\title{percentMatch function}
\usage{
percentMatch(dat, comparerows = TRUE)
}
\arguments{
\item{dat}{dataframe.}

\item{comparerows}{determines if the duplicates or rows or columns needto be calculated (TRUE, by default)}
}
\value{
Returns the list elections and pipe.table objects.
}
\description{
This function calculates the percent match, meaning the maximum percentage of variables that an observation shares with any other observation in the data set.
}
\examples{
library(SurveyQuality)
library(dplyr)
dat <- read.csv(system.file("elites_survey.csv", package="SurveyQuality"))

dat[dat==98|dat==99|dat==95|dat==998|dat==999]<-NA
pmatch<-formatData(dat, rescalev=FALSE, removepercentNAshigher=0.90,
                   cleanfromtxtconstants=TRUE)\%>\%percentMatch()
}
