% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath.R
\name{shortestPath}
\alias{shortestPath}
\title{shortestPath function}
\usage{
shortestPath(dat, vec)
}
\arguments{
\item{dat}{dataframe}

\item{vec}{vector of variables:values belonging to the shortest path}
}
\value{
Returns the list of results: the global shortcuts mean, vector of shortcuts per interviewer, the matrix of shortcuts per interviewer-question (raw), the matrix of shortcuts per interviewer-question (proportion).
}
\description{
This function computes the percentage of cases falling into the shortest path.
}
\examples{
library(SurveyQuality)
library(dplyr)
dat <- read.csv(system.file("elites_survey.csv", package="SurveyQuality"))

dat[dat==98|dat==99|dat==95|dat==998|dat==999]<-NA
shortcut_vars<-c("Q1:1", "Q11:9")
shortpath<-shortestPath(dat, shortcut_vars)
shortpath$var.cuts.prop
}
