% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceAnalysis.R
\name{varianceAnalysis}
\alias{varianceAnalysis}
\title{varianceAnalysis function}
\usage{
varianceAnalysis(dat)
}
\arguments{
\item{dat}{dataframe}
}
\value{
Returns the vector of variances for each interviewer.
}
\description{
This function computes the variance per interviewer.
}
\examples{
library(SurveyQuality)
library(dplyr)
dat <- read.csv(system.file("elites_survey.csv", package="SurveyQuality"))
dat[dat==98|dat==99|dat==95|dat==998|dat==999]<-NA
dat<-dat[,grepl("^Q|A5$", colnames(dat))]

var_test<-formatData(dat, rescalev=FALSE,
                          removepercentNAshigher=0.80,
                          cleanfromtxtconstants=TRUE,
                          keepunscaled="A5")\%>\%
                          varianceAnalysis()
var_test
}
